refactor : 불필요한 import제거 및 변수명 변경, 기타 줄 바꿈 등 변경
