+++ src/main/java/nextstep/courses/domain/session/PaySession.java
@@ -2,7 +2,9 @@ package nextstep.courses.domain.session;
 
 import nextstep.courses.domain.session.coverimage.CoverImages;
 import nextstep.courses.domain.session.enroll.Enrolment;
-import nextstep.courses.domain.session.enroll.RecruitingStatus;
+import nextstep.courses.domain.session.enums.PayType;
+import nextstep.courses.domain.session.enums.RecruitingStatus;
+import nextstep.courses.domain.session.enums.SessionStatus;
 import nextstep.courses.domain.session.period.Period;
 import nextstep.courses.domain.session.student.SessionStudent;
 import nextstep.courses.domain.session.student.SessionStudents;
