+++ src/main/java/nextstep/courses/domain/session/enums/PayType.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain.session;
+package nextstep.courses.domain.session.enums;
 
 public enum PayType {
