+++ src/main/java/nextstep/courses/domain/session/enums/SessionStatus.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain.session;
+package nextstep.courses.domain.session.enums;
 
 public enum SessionStatus {
