+++ src/main/java/nextstep/courses/dto/SelectInfo.java
@@ -1,25 +1,25 @@
 package nextstep.courses.dto;
 
-import nextstep.courses.domain.session.student.SelectionStatus;
+import nextstep.courses.domain.session.enums.SelectionStatus;
 
 public class SelectInfo {
 
-    private Long studentId;
     private Long sessionId;
+    private Long studentId;
     private SelectionStatus selectionStatus;
 
-    public SelectInfo(Long studentId, Long sessionId, SelectionStatus selectionStatus) {
-        this.studentId = studentId;
+    public SelectInfo(Long sessionId, Long studentId, SelectionStatus selectionStatus) {
         this.sessionId = sessionId;
+        this.studentId = studentId;
         this.selectionStatus = selectionStatus;
-    }
-
-    public Long getStudentId() {
-        return studentId;
     }
 
     public Long getSessionId() {
         return this.sessionId;
+    }
+
+    public Long getStudentId() {
+        return studentId;
     }
 
     public SelectionStatus getSelectionStatus() {
