+++ src/main/java/nextstep/courses/service/SessionService.java
@@ -2,7 +2,7 @@ package nextstep.courses.service;
 
 import nextstep.courses.domain.session.student.SessionStudent;
 import nextstep.courses.dto.EnrolmentInfo;
-import nextstep.courses.domain.session.PayType;
+import nextstep.courses.domain.session.enums.PayType;
 import nextstep.courses.dto.SelectInfo;
 
 public interface SessionService {
