+++ src/test/java/nextstep/courses/domain/session/PaySessionTest.java
@@ -1,8 +1,8 @@
 package nextstep.courses.domain.session;
 
 import nextstep.courses.domain.session.coverimage.CoverImages;
-import nextstep.courses.domain.session.enroll.Enrolment;
-import nextstep.courses.domain.session.enroll.RecruitingStatus;
+import nextstep.courses.domain.session.enums.RecruitingStatus;
+import nextstep.courses.domain.session.enums.SessionStatus;
 import nextstep.courses.domain.session.period.Period;
 import nextstep.courses.domain.session.student.SessionStudents;
 import nextstep.courses.dto.EnrolmentInfo;
@@ -11,12 +11,11 @@ import org.junit.jupiter.api.Test;
 
 import java.time.LocalDate;
 
-import static nextstep.courses.domain.session.PayType.*;
-import static nextstep.courses.domain.session.SessionStatus.*;
-import static nextstep.courses.domain.session.enroll.RecruitingStatus.RECRUITING_OFF;
-import static nextstep.courses.domain.session.enroll.RecruitingStatus.RECRUITING_ON;
+import static nextstep.courses.domain.session.enums.PayType.*;
+import static nextstep.courses.domain.session.enums.SessionStatus.*;
+import static nextstep.courses.domain.session.enums.RecruitingStatus.RECRUITING_OFF;
+import static nextstep.courses.domain.session.enums.RecruitingStatus.RECRUITING_ON;
 import static nextstep.users.domain.fixture.DomainFixture.JAVAJIGI;
-import static nextstep.users.domain.fixture.DomainFixture.SANJIGI;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 class PaySessionTest {
@@ -25,7 +24,7 @@ class PaySessionTest {
     @Test
     void validateRecruitingStatus() {
         // given
-        PaySession paySession = createPaySession(PROGRESS, RECRUITING_OFF, 10000L, 2);
+        PaySession paySession = createPaySession(PROGRESS, RECRUITING_OFF);
         EnrolmentInfo enrolmentInfo = createEnrolmentInfo(paySession.id, JAVAJIGI.getId(), 10000L);
 
         // when & then
@@ -37,7 +36,7 @@ class PaySessionTest {
     @Test
     void validateSessionStatus() {
         // given
-        PaySession paySession = createPaySession(PREPARE, RECRUITING_ON, 10000L, 2);
+        PaySession paySession = createPaySession(PREPARE, RECRUITING_ON);
         EnrolmentInfo enrolmentInfo = createEnrolmentInfo(paySession.id, JAVAJIGI.getId(), 10000L);
 
         // when & then
@@ -49,7 +48,7 @@ class PaySessionTest {
     @Test
     void validatePayAmount() {
         // given
-        PaySession paySession = createPaySession(PROGRESS, RECRUITING_ON, 10000L, 2);
+        PaySession paySession = createPaySession(PROGRESS, RECRUITING_ON);
         EnrolmentInfo enrolmentInfo = createEnrolmentInfo(paySession.id, JAVAJIGI.getId(), 12000L);
 
         // when & then
@@ -57,7 +56,7 @@ class PaySessionTest {
             .hasMessage("결제 금액이 강의 금액과 일치하지 않습니다. 강의 금액 :: 10,000원");
     }
 
-    private PaySession createPaySession(SessionStatus sessionStatus, RecruitingStatus recruitingStatus, long amount, int studentCapacity) {
+    private PaySession createPaySession(SessionStatus sessionStatus, RecruitingStatus recruitingStatus) {
         return new PaySession(
             1L,
             PAY,
@@ -69,8 +68,8 @@ class PaySessionTest {
                 LocalDate.of(2023, 12, 5),
                 LocalDate.now()
             ),
-            amount,
-            studentCapacity
+            10000L,
+            2
             );
     }
