+++ src/test/java/nextstep/courses/domain/session/student/SessionStudentsTest.java
@@ -3,7 +3,7 @@ package nextstep.courses.domain.session.student;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import static nextstep.courses.domain.session.student.SelectionStatus.*;
+import static nextstep.courses.domain.session.enums.SelectionStatus.*;
 import static org.assertj.core.api.Assertions.*;
 
 class SessionStudentsTest {
